backend/requirements.txt
django==5.0.6
python-dotenv==1.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
apscheduler==3.10.4
psutil==5.9.8
paho-mqtt>=2.0.0,<3
